// front/src/providers/api-client.generated/model/appValueDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AppTypeDto } from './appTypeDto';
import { TranslationDto } from './translationDto';


export interface AppValueDto { 
    id?: string;
    label: string;
    order?: number;
    code: string;
    appType: AppTypeDto;
    appTypeId: string;
    enabled: boolean;
    translations?: Array<TranslationDto>;
}