// front/src/providers/api-client.generated/model/candidateDto.ts
export interface CandidateDto { 
    id?: string;
    lastName: string;
    firstName: string;
    nickName: string;
    genderId: string;
    gender?: AppValueDto;
    birthDate: Date;
    phone: string;
    phoneSecondary: string;
    email: string;
    additionalEmails?: Array<string>;
    nationality: string;
    relationshipStatusId: string;
    relationshipStatus?: AppValueDto;
    skills: string;
    inCouple: boolean;
    isJobHoused: string;
    hasLicenceDriver: boolean;
    languageId?: string;
    language?: LanguageDto;
    dependentChildren: number;
    animal: boolean;
    isAvailable?: boolean;
    creationDate?: Date;
    modifDate?: Date;
    addresses?: Array<AddressDto>;
    candidateStatusId?: string;
    candidateStatus?: AppValueDto;
    placedJobOfferId?: string;
    placedJobOffer?: JobOfferDto;
    candidateApplications?: Array<CandidateApplicationDto>;
    contractTypeAskedId?: string;
    contractTypeAsked?: AppValueDto;
    workingTimeTypeId?: string;
    workingTimeType?: AppValueDto;
    files?: Array<CandidateFileDto>;
    mainPhotoBase64?: string;
    mainPhotoBase64MimeType?: string;
    note?: number;
    noteItems?: Array<NoteItemDto>;
    candidateJobs?: Array<CandidateJobDto>;
    candidateLicences?: Array<CandidateLicenceDto>;
    interviews?: Array<InterviewDto>;
    candidateLanguages?: Array<CandidateLanguageDto>;
    candidateChildrens?: Array<CandidateChildrenDto>;
    partnerFirstName?: string;
    partnerLastName?: string;
    partnerGenderId?: string;
    partnerGender?: AppValueDto;
    partnerBirthDate?: Date;
    partnerEmail?: string;
    partnerPhone?: string;
    userAlreadyExist?: boolean;
    lastCandidateMessageSendedDate?: Date;
    candidateMessagesUnseen?: boolean;
    candidateAdvancementPercent?: number;
    candidateApplicationsLength?: number;
    hasNoChildren?: boolean;
    candidateReadonlyProperties?: Array<CandidateReadonlyPropertyDto>;
    isOnPost?: boolean;
    consultant?: UserDto;
    consultantId?: string;
    disabled?: boolean;
    jobAdderContactId?: number;
    isVehicle?: boolean;
    candidatePets?: Array<CandidatePetDto>;
    candidateFieldsMiss?: Array<string>;
    candidateContracts?: Array<CandidateContractDto>;
    candidateCountries?: Array<CandidateCountryDto>;
    candidateDepartments?: Array<CandidateDepartmentDto>;
    manuallyCompleted?: boolean;
    mailSentAfterMigration?: boolean;
    associatedUser?: UserDto;
    candidateCurrentJobs?: Array<CandidateCurrentJobDto>;
    globalMobility?: boolean;
    hasManyTravel?: boolean;
    referencesValidated?: boolean;
    allergy?: CandidateAllergiesEnum;
    newsletterUnsubscribed: boolean;
    newsletterUnsubscribedGuid?: string;
    allowed_to_work_us?: boolean;
    require_sponsorship_us?: boolean;
    city?: string;
    jobTitle?: string;
}