// front/src/providers/api-client.generated/model/candidateFileDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AppFileDto } from './appFileDto';
import { CandidateDto } from './candidateDto';
import { AppValueDto } from './appValueDto';


export interface CandidateFileDto { 
    id?: string;
    creationDate?: Date;
    modifDate?: Date;
    fileId?: string;
    file?: AppFileDto;
    candidateId?: string;
    candidate?: CandidateDto;
    fileType?: AppValueDto;
    fileTypeId?: string;
    isMandatory?: boolean;
}