// front/src/providers/api-client.generated/model/getCandidateResponse.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CandidateDto } from './candidateDto';


export interface GetCandidateResponse { 
    success: boolean;
    message?: string;
    error?: object;
    statusCode?: number;
    errorGuid?: string;
    token?: string;
    candidate: CandidateDto;
}